1、全局價格改爲double型 2、好友專業個性簽名和返回異常修復 3、支付列表調整

+++ twant/app/src/main/java/com/ftofs/twant/entity/Footprint.java
@@ -19,7 +19,7 @@ public class Footprint {
 
 
     public Footprint(int footprintId, String date, int storeId, String storeName, int commonId, String imageSrc,
-                     String goodsName, String jingle, float price) {
+                     String goodsName, String jingle, double price) {
         this.footprintId = footprintId;
         this.date = date;
         this.storeId = storeId;
@@ -42,5 +42,5 @@ public class Footprint {
     public String imageSrc;
     public String goodsName;
     public String jingle;
-    public float price;
+    public double price;
 }
