+++ twant/app/src/main/java/com/ftofs/twant/widget/SpecSelectPopup.java
@@ -241,7 +241,7 @@ public class SpecSelectPopup extends BottomPopupView implements View.OnClickList
                             goodsInfo.commonId = commonId;
                             goodsInfo.goodsFullSpecs = goodsInfoVo.getSafeString("goodsFullSpecs");
                             goodsInfo.specValueIds = goodsInfoVo.getSafeString("specValueIds");
-                            goodsInfo.goodsPrice0 = (float) goodsInfoVo.getDouble("goodsPrice0");
+                            goodsInfo.goodsPrice0 = goodsInfoVo.getDouble("goodsPrice0");
                             goodsInfo.price = Util.getSkuPrice(goodsInfoVo);
                             SLog.info("__goodsInfo.price[%s], goodsInfoVo[%s]", goodsInfo.price, goodsInfoVo.toString());
                             goodsInfo.imageSrc = goodsInfoVo.getSafeString("imageSrc");
